+++ src/main/java/live/ioteatime/apiservice/dto/user/BudgetHistoryDto.java
@@ -1,0 +1,17 @@
+package live.ioteatime.apiservice.dto.user;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Getter @Setter
+public class BudgetHistoryDto {
+    int id;
+    @JsonProperty("change_time")
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
+    LocalDateTime changeTime;
+    Long budget;
+}
