+++ src/main/java/live/ioteatime/apiservice/service/AdminService.java
@@ -1,7 +1,7 @@
 package live.ioteatime.apiservice.service;
 
-import live.ioteatime.apiservice.dto.BudgetHistoryDto;
 import live.ioteatime.apiservice.dto.OrganizationDto;
+import live.ioteatime.apiservice.dto.user.BudgetHistoryDto;
 import live.ioteatime.apiservice.dto.user.UserDto;
 
 import java.util.List;
